room: view, update

+++ src/Controller/RoomController.php
@@ -68,21 +68,27 @@ final class RoomController extends AbstractController
     #[Route('/{id}', name: 'app_room_show', methods: ['GET'])]
     public function show(Room $room): Response
     {
+        $form = $this->createForm(RoomType::class, $room);
+        // dump($room);
+        // die();
         return $this->render('room/show.html.twig', [
             'room' => $room,
+            'form' => $form,
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'app_room_edit', methods: ['GET', 'POST'])]
+    #[Route('/{id}/edit', name: 'app_room_edit', methods: ['POST'])]
     public function edit(Request $request, Room $room, EntityManagerInterface $entityManager): Response
     {
+        // dump($request, $room);
+        // die();
         $form = $this->createForm(RoomType::class, $room);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->flush();
 
-            return $this->redirectToRoute('app_room_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
         }
 
         return $this->render('room/edit.html.twig', [
@@ -91,14 +97,27 @@ final class RoomController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_room_delete', methods: ['POST'])]
+    #[Route('/{id}/delete', name: 'app_room_delete', methods: ['GET', 'POST'])]
     public function delete(Request $request, Room $room, EntityManagerInterface $entityManager): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$room->getId(), $request->getPayload()->getString('_token'))) {
-            $entityManager->remove($room);
-            $entityManager->flush();
+        // dump("o xoa roi", $request, $room);
+        // die();
+        if($request->getMethod() == 'POST'){
+            if ($this->isCsrfTokenValid('delete'.$room->getId(), $request->getPayload()->getString('_token'))) {
+                $entityManager->remove($room);
+                $entityManager->flush();
+                return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
+            }
+            else{
+                return $this->redirectToRoute('app_room_delete', [], Response::HTTP_SEE_OTHER);
+            }
+            
         }
+        return $this->render('room/_delete_form.html.twig', [
+            'room' => $room,
+        ]);
+        
 
-        return $this->redirectToRoute('app_room_index', [], Response::HTTP_SEE_OTHER);
+        
     }
 }
